LPS-99312 Move deployDependencies task
